파이썬프로그래밍2주차

+++ CircleCal.py
@@ -1,4 +1,4 @@
-a = int(input("반지름 값을 입력하세요(cm) : "))
+a = float(input("반지름 값을 입력하세요(cm) : "))
 result = a * 2 * 3.1416
 print("원의 둘레 =", result, 'cm')
 result = a ** 2 * 3.1416
